refactor(cpp): split targets/home from nativehook and rename the module lib
- Move the desktop-specific entry point, resolver, layout, hooks and
  motion sources into app/src/main/cpp/targets/home/, keeping nativehook/
  as the target-agnostic hook infrastructure; pure moves, no content
  changes, git rename detection intact
- CMakeLists: point target_sources at the new layout and add a root
  include directory so `nativehook/...` includes stay untouched
- Rename the packaged library to libHyperCeilerNative.so: CMake target,
  the loadLibrary constants in NativeHomeHooksOS4, and the
  META-INF/xposed/native_init.list entry are updated together, since the
  framework matches the list by file name
- Tests: update the relative include paths of the five host tests
