+++ app/src/main/cpp/CMakeLists.txt
@@ -1,21 +1,32 @@
 cmake_minimum_required(VERSION 3.22.1)
-project(hyperceiler_home LANGUAGES CXX)
+project(HyperCeilerNative LANGUAGES CXX)
 
-add_library(hyperceiler_home SHARED hyperceiler_home.cpp)
+# Layout: nativehook/ holds the target-agnostic hook infrastructure; targets/
+# holds the per-feature entry points and resolvers that bind it to one host.
+# The root include dir keeps `nativehook/...` includes valid from any target.
+# The library name is the LSPosed native-hook contract: it must match the
+# `libHyperCeilerNative.so` entry of META-INF/xposed/native_init.list and the
+# `System.loadLibrary` name in NativeHomeHooksOS4.
+add_library(HyperCeilerNative SHARED targets/home/hyperceiler_home.cpp)
 
 option(HYPERCEILER_DOCK_NATIVE_MOTION "Dynamically resolve OS4 native recents animation" ON)
 if(HYPERCEILER_DOCK_NATIVE_MOTION)
     enable_language(ASM)
-    target_sources(hyperceiler_home PRIVATE dock_native_hooks.cpp dock_native_motion.cpp dock_native_motion_arm64.S)
-    target_compile_definitions(hyperceiler_home PRIVATE HYPERCEILER_DOCK_NATIVE_MOTION=1)
+    target_sources(HyperCeilerNative PRIVATE
+        targets/home/dock_native_hooks.cpp
+        targets/home/dock_native_motion.cpp
+        targets/home/dock_native_motion_arm64.S)
+    target_compile_definitions(HyperCeilerNative PRIVATE HYPERCEILER_DOCK_NATIVE_MOTION=1)
 endif()
 
-target_compile_features(hyperceiler_home PRIVATE cxx_std_20)
-target_compile_options(hyperceiler_home PRIVATE
+target_include_directories(HyperCeilerNative PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
+
+target_compile_features(HyperCeilerNative PRIVATE cxx_std_20)
+target_compile_options(HyperCeilerNative PRIVATE
     -Wall
     -Wextra
     -Werror
     -fvisibility=hidden
 )
 
-target_link_libraries(hyperceiler_home PRIVATE android binder_ndk log)
+target_link_libraries(HyperCeilerNative PRIVATE android binder_ndk log)
