+++ app/src/main/cpp/targets/home/dock_native_motion_arm64.S
