+++ library/libhook/src/main/resources/META-INF/xposed/native_init.list
@@ -1,1 +1,1 @@
-libhyperceiler_home.so
+libHyperCeilerNative.so
