+++ tests/home-dock-window/DockNativeResolverTest.cpp
@@ -1,5 +1,5 @@
 /* SPDX-License-Identifier: AGPL-3.0-or-later */
-#include "../../app/src/main/cpp/dock_native_resolver.h"
+#include "../../app/src/main/cpp/targets/home/dock_native_resolver.h"
 #include <cassert>
 #include <cstring>
 #include <fstream>
